MBusDataFixed: Add enumerated properties.
This adds some fields to decode the fixed frame medium and unit types.
It is modelled on mbus_data_fixed_unit and mbus_data_fixed_medium
functions in libmbus (not exposed by python-mbus).

The enumerations make use of the `enum` module: part of the standard
Python 3 library, or using the `enum34` module in Python 2.x.