MBus: Fix arguments/return types of values.
It seems `ctypes` in Python assumes that arguments that look like
integers, are 32-bit integers.  Including things like pointers.

This causes all hell to break loose on an AMD64 host (probably on any
other 64-bit host too) if the pointers happen to get allocated above the
4GB address space boundary.

https://bugs.gentoo.org/show_bug.cgi?id=567418

The fix is to correctly assign the data types for each argument and
return type in each C function call.